devop: move token balances under balance module
